Add query fix to remaining effects
