+++ src/app/store/it-system-interfaces/effects.ts
@@ -7,7 +7,7 @@ import { compact } from 'lodash';
 import { catchError, combineLatestWith, map, of, switchMap } from 'rxjs';
 import { APIV2ItInterfaceService } from 'src/app/api/v2';
 import { INTERFACE_COLUMNS_ID } from 'src/app/shared/constants/persistent-state-constants';
-import { replaceQueryByMultiplePropertyContains } from 'src/app/shared/helpers/odata-query.helpers';
+import { replaceDuplicateRangeVariables, replaceQueryByMultiplePropertyContains } from 'src/app/shared/helpers/odata-query.helpers';
 import { toODataString } from 'src/app/shared/models/grid-state.model';
 import { adaptITInterface } from 'src/app/shared/models/it-interface/it-interface.model';
 import { OData } from 'src/app/shared/models/odata.model';
@@ -231,5 +231,5 @@ function applyQueryFixes(odataString: string): string {
     lastChangedByUserSearchedProperties
   );
 
-  return fixedOdataString;
+  return replaceDuplicateRangeVariables(fixedOdataString);
 }
