+++ src/app/store/organization/organization-user/effects.ts
@@ -14,6 +14,7 @@ import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
 import { GridColumnStorageService } from 'src/app/shared/services/grid-column-storage-service';
 import { selectOrganizationUuid } from '../../user-store/selectors';
 import { OrganizationUserActions } from './actions';
+import { replaceDuplicateRangeVariables } from 'src/app/shared/helpers/odata-query.helpers';
 
 @Injectable()
 export class OrganizationUserEffects {
@@ -222,9 +223,10 @@ function applyQueryFixes(odataString: string) {
   const objectOwnerColumn = 'ObjectOwner.Name';
   const nameColumn = 'Name';
 
-  return odataString
+  odataString = odataString
     .replace(getNameFilterPattern(objectOwnerColumn), `$1concat(concat(ObjectOwner/Name, ' '), ObjectOwner/LastName)$2`)
     .replace(getNameFilterPattern(nameColumn), `$1concat(concat(Name, ' '), LastName)$2`);
+  return replaceDuplicateRangeVariables(odataString);
 }
 
 function getNameFilterPattern(column: string) {
